Basic console functions
MUST DOCUMENT EVERYTHING DONE SO FAR!! I will NOT make the mistake I did with the other simulator, I have to think this through to the end before writing too much code. It feels good to just hack and solve problems, but I tend to wall myself in when I do that so tomorrow I'm going to paper to make sure the right hooks are in place for extendability.  Proud of the encapsulation in the Output object.