updated my knowledge of JS encapsulation
Finally understood more clearly how to encapsulate a member of an object. declaring _output_text in the parameter list of the object was not encapsulating it. Not sure how I ever felt it was. I understand closure somewhat better now and recognize that much info online could be improved to say the least. Renamed Output.send_output method to write_output. Rewrote the mothod for clarity as I noticed that some of its functionality fell into a category of its own. It is now neatly divided into three areas of concern: 1. moving the active line down, 2. updating the contents of the screen_buffer, and 3. reinitializing the terminal to allow further input.