split up JS file, unidioted some comments
Split terminal and Output into separate files.  uncovered some flawed thinking regarding the prompt text. for some reason I had it in my head that the declarations of the variables that compose the prompt needed to be near the bottom of the JS since the spans were dynamically generated. It just landed on me that though these variables contain HTML markup that gets injected into the active_line's .innerHTML attribute and generates an element id="al" <-'active_line', the variables are just strings, and can go at the top with the rest of the declarations. :neutral_face: