refactoring class structure
As I'm learning more about JS, I'm starting to see more things from a progressively more OOP mindset. I'm in the process of separating the concerns of Terminal and Output into the more clear relationship of the Terminal and its Buffer. I've created the 'net-sim-refactor' folder to hold this code until it is fully built.